requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
djangorestframework==3.14.0
requests==2.28.1
geopy==2.3.0
